+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Drag to reorder signals now responds to scroll
   - Surfer surver support
   - Status Bar now shows value change count for selected signal
+  - Automatically reload waveform on file updates
 - Fixed
   - Renderer floating point math is much more accurate resulting in less jitter
   - Revealing signals in viewport scrolls to the signal in question